fix(quest-06)
