+++ solutions/scytale_cipher/src/lib.rs
@@ -1,21 +1,23 @@
-pub fn scytale_cipher(s: String, i: u32) -> String {
-    if i as usize >= s.chars().count() || i == 1 {
-        return s.to_string();
+use std::iter;
+
+pub fn scytale_cipher(message: &str, i: usize) -> String {
+    let i = i.min(message.len());
+    if message.is_empty() || i == 0 {
+        String::new()
+    } else {
+        let cols = message.len().div_ceil(i);
+
+        (0..i)
+            .map(|n| {
+                message
+                    .chars()
+                    .skip(n)
+                    .step_by(i)
+                    .chain(iter::repeat_n(' ', cols - (message.len() - n).div_ceil(i)))
+                    .collect::<String>()
+            })
+            .collect::<String>()
+            .trim_end()
+            .to_owned()
     }
-
-    let width = (s.chars().count() as f64 / i as f64).ceil() as usize;
-    let mut table = vec![vec![' '; width]; i as usize];
-
-    for (pos, element) in s.chars().enumerate() {
-        let col = pos % i as usize;
-        let row = pos / i as usize;
-
-        table[col][row] = element;
-    }
-    table
-        .iter()
-        .flatten()
-        .collect::<String>()
-        .trim_end()
-        .to_string()
 }
