+++ tests/spelling_test/Cargo.toml
@@ -7,5 +7,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rand = "0.7"
 spelling = { path = "../../solutions/spelling"}
