--- tests/stars_test/src/main.rs
@@ -1,41 +1,0 @@
-/*
-## stars
-
-### Instructions
-
-Write a recursive function named `stars` that takes an i32 as
-parameter and returns a String of stars (asterisks) 2^n long (2 to the nth power).
-
-### Example:
-
-stars(0)     "*"                     2^0 = 1
-stars(1)     "**"                    2^1 = 2
-stars(2)     "****"                  2^2 = 4
-stars(3)     "********"              2^3 = 8
-stars(4)     "****************"      2^4 = 16
-
-*/
-use stars::*;
-
-fn main() {
-    println!("{}", stars(4));
-    println!("{}", stars(5));
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    #[test]
-    fn test_stars() {
-        assert_eq!(stars(0), "*");
-        assert_eq!(stars(1), "**");
-        assert_eq!(stars(2), "****");
-        assert_eq!(stars(3), "********");
-        assert_eq!(stars(4), "****************");
-        assert_eq!(stars(5), "********************************");
-        assert_eq!(
-            stars(10),
-            "****************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************************"
-        );
-    }
-}
